requirements.txt
requests
rich
gitpython
cryptography
python-dotenv
